+++ src/rawmessage.rs
@@ -17,8 +17,8 @@ use std;
 pub trait NoPointers { }
 
 struct Internal {
-    len:            uint,
-    cap:            uint,
+    len:            usize,
+    cap:            usize,
     buf:            *mut u8,
 }
 
@@ -26,12 +26,12 @@ unsafe impl Send for Internal{}
 
 impl Drop for Internal {
     fn drop(&mut self) {
-        unsafe { deallocate(self.buf, self.cap, size_of::<uint>()); }
+        unsafe { deallocate(self.buf, self.cap, size_of::<usize>()); }
     }
 }
 
 impl Internal {
-    fn new(mut cap: uint) -> Internal {
+    fn new(mut cap: usize) -> Internal {
         // This helps let sync and clone messages work. Since sometimes
         // they may use a zero size type just for conveying a signal of
         // some sort and the will try to allocate a message of zero bytes
@@ -44,11 +44,11 @@ impl Internal {
         Internal {
             len:        cap,
             cap:        cap,
-            buf:        unsafe { allocate(cap, size_of::<uint>()) },
+            buf:        unsafe { allocate(cap, size_of::<usize>()) },
         }
     }
 
-    fn id(&self) -> uint {
+    fn id(&self) -> usize {
         unsafe { transmute(self) }
     }
 
@@ -56,7 +56,7 @@ impl Internal {
         let i = Internal {
             len:        self.cap,
             cap:        self.cap,
-            buf:        unsafe { allocate(self.cap, size_of::<uint>() ) },
+            buf:        unsafe { allocate(self.cap, size_of::<usize>() ) },
         };
 
         unsafe { 
@@ -66,7 +66,7 @@ impl Internal {
         i
     }
 
-    fn setlen(&mut self, len: uint) {
+    fn setlen(&mut self, len: usize) {
         if len > self.cap {
             panic!("len:{} must be <= cap:{}", len, self.cap);
         }
@@ -74,9 +74,9 @@ impl Internal {
         self.len = len;
     }
 
-    fn resize(&mut self, newcap: uint) {
+    fn resize(&mut self, newcap: usize) {
         unsafe {
-            let nbuf = allocate(newcap, size_of::<uint>());
+            let nbuf = allocate(newcap, size_of::<usize>());
 
             if newcap <= self.cap {
                 copy_memory(nbuf, self.buf, self.cap);
@@ -84,7 +84,7 @@ impl Internal {
                 copy_memory(nbuf, self.buf, newcap);
             }
 
-            deallocate(self.buf, self.cap, size_of::<uint>());
+            deallocate(self.buf, self.cap, size_of::<usize>());
             self.buf = nbuf;
 
             self.cap = newcap;
@@ -113,7 +113,7 @@ impl Clone for RawMessage {
 
 impl RawMessage {
     /// Create a new raw message with the specified capacity.
-    pub fn new(cap: uint) -> RawMessage {
+    pub fn new(cap: usize) -> RawMessage {
         RawMessage {i:  Arc::new(Mutex::new(Internal::new(cap)))}
     }
 
@@ -124,7 +124,7 @@ impl RawMessage {
         rm
     }
 
-    pub fn id(&self) -> uint {
+    pub fn id(&self) -> usize {
         unsafe { transmute(self) }
     }
 
@@ -134,18 +134,18 @@ impl RawMessage {
     pub fn new_fromstr(s: &str) -> RawMessage {
         let m = RawMessage::new(s.len());
         unsafe {
-            copy_memory(m.i.lock().unwrap().buf, *(transmute::<&&str, *const uint>(&s)) as *const u8, s.len());
+            copy_memory(m.i.lock().unwrap().buf, *(transmute::<&&str, *const usize>(&s)) as *const u8, s.len());
         }
         m
     }
 
     /// Get the capacity.
-    pub fn cap(&self) -> uint {
+    pub fn cap(&self) -> usize {
         /*unsafe {
             let m: &Mutex<Internal> = &*self.i;
-            let ptr: *mut uint = transmute(&*self.i);
-            let ptr: uint = *ptr;
-            *((ptr + 8 * 2) as *mut uint) = 0;
+            let ptr: *mut usize = transmute(&*self.i);
+            let ptr: usize = *ptr;
+            *((ptr + 8 * 2) as *mut usize) = 0;
             std::io::stdio::stdout_raw().write(
             format!("@trying to lock\nlock:{:x}\ncount:{:x}\nowner:{:x}\nnusers:{:x}\n",
                 *((ptr + 4 * 0) as *mut u32),
@@ -157,8 +157,8 @@ impl RawMessage {
         let i = self.i.lock().unwrap();
         /*unsafe {
             let m: &Mutex<Internal> = &*self.i;
-            let ptr: *mut uint = transmute(&*self.i);
-            let ptr: uint = *ptr;
+            let ptr: *mut usize = transmute(&*self.i);
+            let ptr: usize = *ptr;
             std::io::stdio::stdout_raw().write(
             format!("@locked\nlock:{:x}\ncount:{:x}\nowner:{:x}\nnusers:{:x}\n",
                 *((ptr + 4 * 0) as *mut u32),
@@ -177,22 +177,22 @@ impl RawMessage {
     /// Set the length.
     ///
     /// The length must not exceed the capacity or it will panic.
-    pub fn setlen(&mut self, len: uint) {
+    pub fn setlen(&mut self, len: usize) {
         self.i.lock().unwrap().setlen(len);
     }
 
     /// Get the length.
-    pub fn len(&self) -> uint {
+    pub fn len(&self) -> usize {
         self.i.lock().unwrap().len
     }
 
     /// Resize the capacity of the message keeping the old contents or truncating them.
-    pub fn resize(&mut self, newcap: uint) {
+    pub fn resize(&mut self, newcap: usize) {
         self.i.lock().unwrap().resize(newcap);
     }
 
     /// Write into the buffer from a byte slice.
-    pub fn write_from_slice(&mut self, mut offset: uint, f: &[u8]) {
+    pub fn write_from_slice(&mut self, mut offset: usize, f: &[u8]) {
         let mut i = self.i.lock().unwrap();
 
         if offset + f.len() > i.cap {
@@ -200,7 +200,7 @@ impl RawMessage {
         }
 
         for item in f.iter() {
-            unsafe { *((i.buf as uint + offset) as *mut u8) = *item };
+            unsafe { *((i.buf as usize + offset) as *mut u8) = *item };
             offset += 1;
         }
 
@@ -228,18 +228,18 @@ impl RawMessage {
     /// Safely write a structure/value by reference into the buffer at the specified offet.
     ///
     /// The offset must not exceed the capacity. The length is not updated.
-    pub fn writestructref<T>(&mut self, offset: uint, t: &T) {
+    pub fn writestructref<T>(&mut self, offset: usize, t: &T) {
         let i = self.i.lock().unwrap();
 
         if offset + size_of::<T>() > i.cap {
             panic!("write past end of buffer! {}:{}", offset + size_of::<T>(), i.cap);
         }
 
-        unsafe { copy_memory::<T>((i.buf as uint + offset) as *mut T, transmute(t), 1); }        
+        unsafe { copy_memory::<T>((i.buf as usize + offset) as *mut T, transmute(t), 1); }        
     }
 
     /// Safely write a structure/value by consuming it.
-    pub fn writestruct<T>(&mut self, offset: uint, t: T) {
+    pub fn writestruct<T>(&mut self, offset: usize, t: T) {
         self.writestructref(offset, &t);
     }
 
@@ -248,55 +248,55 @@ impl RawMessage {
     /// This function is only intended to help keep you from having to use
     /// unsafe code blocks. You can easily mark a structure as safe and
     /// crash your program.
-    pub fn readstruct<T: NoPointers>(&self, offset: uint) -> T {
+    pub fn readstruct<T: NoPointers>(&self, offset: usize) -> T {
         unsafe { self.readstructunsafe(offset) }
     }
 
     /// Read a structure only if it is marked as safe into a mutable reference.
-    pub fn readstructref<T: NoPointers>(&self, offset: uint, t: &mut T) {
+    pub fn readstructref<T: NoPointers>(&self, offset: usize, t: &mut T) {
         unsafe { self.readstructunsaferef(offset, t) };
     }
 
     /// This is the same as readstruct except it is an unsafe call.
-    pub unsafe fn readstructunsafe<T>(&self, offset: uint) -> T {
+    pub unsafe fn readstructunsafe<T>(&self, offset: usize) -> T {
         let mut out: T = uninitialized::<T>();
         self.readstructunsaferef(offset, &mut out);
         out
     }
 
     /// This is the same as readstructref except it is an unsafe call.
-    pub unsafe fn readstructunsaferef<T>(&self, offset: uint, t: &mut T) {
+    pub unsafe fn readstructunsaferef<T>(&self, offset: usize, t: &mut T) {
         let i = self.i.lock().unwrap();
 
         if offset + size_of::<T>() > i.cap {
             panic!("read past end of buffer!")
         }
 
-        copy_memory(t, (i.buf as uint + offset) as *const T, 1);
+        copy_memory(t, (i.buf as usize + offset) as *const T, 1);
     }
 
     /// Write a unsigned 8-bit value at the specified offset.
-    pub fn writeu8(&mut self, offset: uint, value: u8) { self.writestruct(offset, value); }
+    pub fn writeu8(&mut self, offset: usize, value: u8) { self.writestruct(offset, value); }
     /// Write a unsigned 16-bit value at the specified offset.
-    pub fn writeu16(&mut self, offset: uint, value: u16) { self.writestruct(offset, value); }
+    pub fn writeu16(&mut self, offset: usize, value: u16) { self.writestruct(offset, value); }
     /// Write a unsigned 32-bit value at the specified offset.
-    pub fn writeu32(&mut self, offset: uint, value: u32) { self.writestruct(offset, value); }
+    pub fn writeu32(&mut self, offset: usize, value: u32) { self.writestruct(offset, value); }
     /// Write a signed 8-bit value at the specified offset.
-    pub fn writei8(&mut self, offset: uint, value: i8) { self.writestruct(offset, value); }
+    pub fn writei8(&mut self, offset: usize, value: i8) { self.writestruct(offset, value); }
     /// Write a signed 16-bit value at the specified offset.
-    pub fn writei16(&mut self, offset: uint, value: i16) { self.writestruct(offset, value); }
+    pub fn writei16(&mut self, offset: usize, value: i16) { self.writestruct(offset, value); }
     /// Write a signed 32-bit value at the specified offset.
-    pub fn writei32(&mut self, offset: uint, value: i32) { self.writestruct(offset, value); }
+    pub fn writei32(&mut self, offset: usize, value: i32) { self.writestruct(offset, value); }
     /// Read a unsigned 8-bit value at the specified offset.
-    pub fn readu8(&self, offset: uint) -> u8 { unsafe { self.readstructunsafe(offset) } }
+    pub fn readu8(&self, offset: usize) -> u8 { unsafe { self.readstructunsafe(offset) } }
     /// Read a unsigned 16-bit value at the specified offset.
-    pub fn readu16(&self, offset: uint) -> u16 { unsafe { self.readstructunsafe(offset) } }
+    pub fn readu16(&self, offset: usize) -> u16 { unsafe { self.readstructunsafe(offset) } }
     /// Read a unsigned 32-bit value at the specified offset.
-    pub fn readu32(&self, offset: uint) -> u32 { unsafe { self.readstructunsafe(offset) } }
+    pub fn readu32(&self, offset: usize) -> u32 { unsafe { self.readstructunsafe(offset) } }
     /// Read a signed 8-bit value at the specified offset.
-    pub fn readi8(&self, offset: uint) -> i8 { unsafe { self.readstructunsafe(offset) } }
+    pub fn readi8(&self, offset: usize) -> i8 { unsafe { self.readstructunsafe(offset) } }
     /// Read a signed 16-bit value at the specified offset.
-    pub fn readi16(&self, offset: uint) -> i16 { unsafe { self.readstructunsafe(offset) } }
+    pub fn readi16(&self, offset: usize) -> i16 { unsafe { self.readstructunsafe(offset) } }
     /// Read a signed 32-bit value at the specified offset.
-    pub fn readi32(&self, offset: uint) -> i32 { unsafe { self.readstructunsafe(offset) } }
+    pub fn readi32(&self, offset: usize) -> i32 { unsafe { self.readstructunsafe(offset) } }
 }
